requirements.txt
mysql-connector-python==8.0.33
PyYAML==6.0.1
fastapi
uvicorn
pydantic